requirements.txt
astroid==2.2.5
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-debug-toolbar==1.10.1
django-js-asset==1.2.2
filelock==3.0.12
idna==2.8
importlib-metadata==0.20
isort==4.3.20
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.0
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.16.4
oauthlib==3.1.0
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
selenium==3.141.0
six==1.12.0
sqlparse==0.2.4
toml==0.10.0
tox==3.13.2
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.2
wrapt==1.11.2
zipp==0.6.0
